py/bin/requirements.txt
flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
